requirements.txt
absl-py==1.1.0
accelerate==0.11.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
blobfile==2.1.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.13
docker-pycreds==0.4.0
# Editable install with no version control (drr-projector-function==0.0.0)
-e f:\programdata\anaconda3\envs\medicaldiffusion\lib\site-packages\drr_projector_function-0.0.0-py3.8-win-amd64.egg
easydict==1.11
einops==0.4.1
einops-exts==0.0.3
ema-pytorch==0.0.8
executing==2.0.0
filelock==3.13.1
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
humanize==4.2.2
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
joblib==1.1.0
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.3
lpips==0.1.4
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.7.4
matplotlib-inline==0.1.6
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
monai==1.3.0
mpi4py==3.1.5
multidict==6.0.2
nest-asyncio==1.5.8
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
ninja==1.11.1.1
numpy==1.24.4
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.8.1.78
packaging==23.2
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
platformdirs==3.11.0
prefetch-generator==1.0.3
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.19.0
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.19.0
pyDeprecate==0.3.1
pydicom==2.3.0
Pygments==2.16.1
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.1
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.1
PyWavelets==1.3.0
pywin32==306
PyYAML==6.0
pyzmq==25.1.1
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.1.5
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
threadpoolctl==3.1.0
tifffile==2022.8.3
toml==0.10.2
torch==1.12.0
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchio==0.18.80
torchmetrics==0.9.1
torchvision==0.13.0
tornado==6.3.3
tqdm==4.64.0
traitlets==5.11.2
typing_extensions==4.2.0
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
wandb==0.12.21
wcwidth==0.2.8
Werkzeug==2.1.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
